Deleted access renewal component and distributed pieces of it to other components.
